testing_requirements.txt
aiodns==2.0.0
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
# Editable Git install with no remote (calam-voyager==0.0.1)
-e /home/ben/Projects/calam-voyager
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
idna==2.10
lxml==4.6.2
multidict==5.1.0
numpy==1.19.5
pandas==1.2.1
pycares==3.1.1
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
